#87 Fix check_quote.c for norminette
